third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEColorMatrixElement-svgdom-in-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' property of the SVGFEColorMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS matrixElement.in1.baseVal is "SourceAlpha"
PASS saturateElement.in1.baseVal is "SourceAlpha"
PASS hueRotateElement.in1.baseVal is "SourceAlpha"
PASS luminanceToAlphaElement.in1.baseVal is "SourceAlpha"
PASS matrixElement.in1.baseVal is "SourceGraphic"
PASS saturateElement.in1.baseVal is "SourceGraphic"
PASS hueRotateElement.in1.baseVal is "SourceGraphic"
PASS luminanceToAlphaElement.in1.baseVal is "SourceGraphic"
PASS successfullyParsed is true

TEST COMPLETE
